+++ modules/security.py
@@ -1,3 +1,3 @@
 """Impostazioni di accesso della PAS Demo."""
 
-DEMO_PASSWORD = "PROVASOFTWARE"
+DEMO_PASSWORD = "PAS2026"